requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
billiard==4.2.0 
certifi==2024.6.2
frozenlist==1.4.1
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
socksio==1.0.0
netaddr==1.3.0
orjson==3.10.3
outcome==1.3.0.post0
pyOpenSSL==24.2.1
cryptography==43.0.0
curl_cffi==0.9.0
ndg-httpsclient==0.5.1
packaging==24.1
prompt_toolkit==3.0.47
pycryptodome==3.20.0
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytunnel==2.4.2
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
setuptools==71.0.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
trio==0.25.1
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
celery==5.4.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==14.1
web3==6.20.0
aiogram==2.23.1
motor==3.5.1
redis==4.5.4
pymongo==4.7.3
dedust==1.1.4